split apps.py to core.py and factory.py
Re-organized the cou/apps to have factory.py, core.py, auxiliary.py,
subordinate.py and auxiliary_subordinate.py.
