+++ tests/unit/apps/test_app.py
@@ -16,8 +16,8 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from cou.apps import app as app_module
-from cou.apps.app import OpenStackApplication
+from cou.apps import factory
+from cou.apps.core import OpenStackApplication
 from cou.exceptions import (
     ApplicationError,
     HaltUpgradePlanGeneration,
@@ -629,10 +629,10 @@ def test_upgrade_plan_application_already_disable_action_managed(status, config,
     assert upgrade_plan == expected_plan
 
 
-@patch.object(app_module, "is_charm_supported", return_value=False)
+@patch.object(factory, "is_charm_supported", return_value=False)
 def test_app_factory_not_supported_openstack_charm(mock_is_charm_supported):
     charm = "my-app"
-    my_app = app_module.AppFactory.create(
+    my_app = factory.AppFactory.create(
         name=charm,
         status=MagicMock(),
         config=MagicMock(),
@@ -643,17 +643,17 @@ def test_app_factory_not_supported_openstack_charm(mock_is_charm_supported):
     mock_is_charm_supported.assert_called_once_with(charm)
 
 
-@patch.object(app_module, "is_charm_supported", return_value=True)
+@patch.object(factory, "is_charm_supported", return_value=True)
 def test_app_factory_register(mock_is_charm_supported):
     charm = "foo"
 
-    @app_module.AppFactory.register_application([charm])
+    @factory.AppFactory.register_application([charm])
     class Foo:
         def __init__(self, *_, **__):
             pass
 
-    assert charm in app_module.AppFactory.charms
-    foo = app_module.AppFactory.create("my-foo", MagicMock(), {}, MagicMock(), charm)
+    assert charm in factory.AppFactory.charms
+    foo = factory.AppFactory.create("my-foo", MagicMock(), {}, MagicMock(), charm)
     mock_is_charm_supported.assert_called_once_with(charm)
     assert foo is not None
     assert isinstance(foo, Foo)
