+++ tests/unit/steps/test_steps_plan.py
@@ -16,7 +16,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from cou.apps.app import OpenStackApplication
+from cou.apps.core import OpenStackApplication
 from cou.exceptions import HaltUpgradePlanGeneration, NoTargetError
 from cou.steps import UpgradeStep
 from cou.steps.analyze import Analysis
